+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Updated for Classic Era patch 1.15.2
 * Fixed Death Knight behavior in Cataclysm Classic
 * Added shortcut to options panel from tab right-click menu
+* Added /engraved command to lock/unlock addon
 
 ### v2.1.2
 * Fixed PLAYER_TALENT_UPDATE error in Classic Era